requirement.txt
streamlit
pillow
opencv-python-headless
rembg
torch
torchvision
diffusers
transformers
accelerate